backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
openai==1.3.5
pydantic>=1.8.0
pydantic-settings==2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
click>=8.0.0
psycopg2-binary>=2.9.0  # For PostgreSQL support
alembic>=1.7.0  # For database migrations 